Crate rusoto_license_manager

AWS License Manager
This is the AWS License Manager API Reference. It provides descriptions, syntax, and usage examples for each of the actions and data types for License Manager. The topic for each action shows the Query API request parameters and the XML response. You can also view the XML request elements in the WSDL.

Alternatively, you can use one of the AWS SDKs to access an API that's tailored to the programming language or platform that you're using. For more information, see AWS SDKs.

If you're using the service, you're probably looking for LicenseManagerClient and LicenseManager.

Structs

ConsumedLicenseSummary
Details about license consumption.

CreateLicenseConfigurationRequest

CreateLicenseConfigurationResponse

DeleteLicenseConfigurationRequest

DeleteLicenseConfigurationResponse

Filter
A filter name and value pair that is used to return a more specific list of results from a describe operation. Filters can be used to match a set of resources by specific criteria, such as tags, attributes, or IDs. The filters supported by a Describe operation are documented with the Describe operation.

GetLicenseConfigurationRequest

GetLicenseConfigurationResponse

GetServiceSettingsRequest

GetServiceSettingsResponse

InventoryFilter
An inventory filter object.

LicenseConfiguration
A license configuration is an abstraction of a customer license agreement that can be consumed and enforced by License Manager. Components include specifications for the license type (licensing by instance, socket, CPU, or VCPU), tenancy (shared tenancy, Amazon EC2 Dedicated Instance, Amazon EC2 Dedicated Host, or any of these), host affinity (how long a VM must be associated with a host), the number of licenses purchased and used.

LicenseConfigurationAssociation
Describes a server resource that is associated with a license configuration.

LicenseConfigurationUsage
Contains details of the usage of each resource from the license pool.

LicenseManagerClient
A client for the AWS License Manager API.

LicenseSpecification
Object used for associating a license configuration with a resource.

ListAssociationsForLicenseConfigurationRequest

ListAssociationsForLicenseConfigurationResponse

ListLicenseConfigurationsRequest

ListLicenseConfigurationsResponse

ListLicenseSpecificationsForResourceRequest

ListLicenseSpecificationsForResourceResponse

ListResourceInventoryRequest

ListResourceInventoryResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

ListUsageForLicenseConfigurationRequest

ListUsageForLicenseConfigurationResponse

ManagedResourceSummary
Summary for a resource.

OrganizationConfiguration
Object containing configuration information for AWS Organizations.

ResourceInventory
A set of attributes that describe a resource.

Tag
Tag for a resource in a key-value format.

TagResourceRequest

TagResourceResponse

UntagResourceRequest

UntagResourceResponse

UpdateLicenseConfigurationRequest

UpdateLicenseConfigurationResponse

UpdateLicenseSpecificationsForResourceRequest

UpdateLicenseSpecificationsForResourceResponse

UpdateServiceSettingsRequest

UpdateServiceSettingsResponse

Enums

CreateLicenseConfigurationError
Errors returned by CreateLicenseConfiguration

DeleteLicenseConfigurationError
Errors returned by DeleteLicenseConfiguration

GetLicenseConfigurationError
Errors returned by GetLicenseConfiguration

GetServiceSettingsError
Errors returned by GetServiceSettings

ListAssociationsForLicenseConfigurationError
Errors returned by ListAssociationsForLicenseConfiguration

ListLicenseConfigurationsError
Errors returned by ListLicenseConfigurations

ListLicenseSpecificationsForResourceError
Errors returned by ListLicenseSpecificationsForResource

ListResourceInventoryError
Errors returned by ListResourceInventory

ListTagsForResourceError
Errors returned by ListTagsForResource

ListUsageForLicenseConfigurationError
Errors returned by ListUsageForLicenseConfiguration

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateLicenseConfigurationError
Errors returned by UpdateLicenseConfiguration

UpdateLicenseSpecificationsForResourceError
Errors returned by UpdateLicenseSpecificationsForResource

UpdateServiceSettingsError
Errors returned by UpdateServiceSettings

Traits

LicenseManager
Trait representing the capabilities of the AWS License Manager API. AWS License Manager clients implement this trait.