Struct rusoto_license_manager::LicenseManagerClient

pub struct LicenseManagerClient { /* fields omitted */ }

A client for the AWS License Manager API.

Methods

impl LicenseManagerClient

pub fn new(region: Region) -> LicenseManagerClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> LicenseManagerClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,