Trait rusoto_license_manager::LicenseManager

pub trait LicenseManager {
    fn create_license_configuration(
        &self, 
        input: CreateLicenseConfigurationRequest
    ) -> RusotoFuture<CreateLicenseConfigurationResponse, CreateLicenseConfigurationError>;

    fn delete_license_configuration(
        &self, 
        input: DeleteLicenseConfigurationRequest
    ) -> RusotoFuture<DeleteLicenseConfigurationResponse, DeleteLicenseConfigurationError>;

    fn get_license_configuration(
        &self, 
        input: GetLicenseConfigurationRequest
    ) -> RusotoFuture<GetLicenseConfigurationResponse, GetLicenseConfigurationError>;

    fn get_service_settings(
        &self
    ) -> RusotoFuture<GetServiceSettingsResponse, GetServiceSettingsError>;

    fn list_associations_for_license_configuration(
        &self, 
        input: ListAssociationsForLicenseConfigurationRequest
    ) -> RusotoFuture<ListAssociationsForLicenseConfigurationResponse, ListAssociationsForLicenseConfigurationError>;

    fn list_license_configurations(
        &self, 
        input: ListLicenseConfigurationsRequest
    ) -> RusotoFuture<ListLicenseConfigurationsResponse, ListLicenseConfigurationsError>;

    fn list_license_specifications_for_resource(
        &self, 
        input: ListLicenseSpecificationsForResourceRequest
    ) -> RusotoFuture<ListLicenseSpecificationsForResourceResponse, ListLicenseSpecificationsForResourceError>;

    fn list_resource_inventory(
        &self, 
        input: ListResourceInventoryRequest
    ) -> RusotoFuture<ListResourceInventoryResponse, ListResourceInventoryError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn list_usage_for_license_configuration(
        &self, 
        input: ListUsageForLicenseConfigurationRequest
    ) -> RusotoFuture<ListUsageForLicenseConfigurationResponse, ListUsageForLicenseConfigurationError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<TagResourceResponse, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<UntagResourceResponse, UntagResourceError>;

    fn update_license_configuration(
        &self, 
        input: UpdateLicenseConfigurationRequest
    ) -> RusotoFuture<UpdateLicenseConfigurationResponse, UpdateLicenseConfigurationError>;

    fn update_license_specifications_for_resource(
        &self, 
        input: UpdateLicenseSpecificationsForResourceRequest
    ) -> RusotoFuture<UpdateLicenseSpecificationsForResourceResponse, UpdateLicenseSpecificationsForResourceError>;

    fn update_service_settings(
        &self, 
        input: UpdateServiceSettingsRequest
    ) -> RusotoFuture<UpdateServiceSettingsResponse, UpdateServiceSettingsError>;
}

Trait representing the capabilities of the AWS License Manager API. AWS License Manager clients implement this trait.

Required methods

fn create_license_configuration(
&self,
input: CreateLicenseConfigurationRequest
) -> RusotoFuture<CreateLicenseConfigurationResponse, CreateLicenseConfigurationError>

Creates a new license configuration object. A license configuration is an abstraction of a customer license agreement that can be consumed and enforced by License Manager. Components include specifications for the license type (licensing by instance, socket, CPU, or VCPU), tenancy (shared tenancy, Amazon EC2 Dedicated Instance, Amazon EC2 Dedicated Host, or any of these), host affinity (how long a VM must be associated with a host), the number of licenses purchased and used.

fn delete_license_configuration(
&self,
input: DeleteLicenseConfigurationRequest
) -> RusotoFuture<DeleteLicenseConfigurationResponse, DeleteLicenseConfigurationError>

Deletes an existing license configuration. This action fails if the configuration is in use.

fn get_license_configuration(
&self,
input: GetLicenseConfigurationRequest
) -> RusotoFuture<GetLicenseConfigurationResponse, GetLicenseConfigurationError>

Returns a detailed description of a license configuration.

fn get_service_settings(
&self
) -> RusotoFuture<GetServiceSettingsResponse, GetServiceSettingsError>

Gets License Manager settings for a region. Exposes the configured S3 bucket, SNS topic, etc., for inspection.

fn list_associations_for_license_configuration(
&self,
input: ListAssociationsForLicenseConfigurationRequest
) -> RusotoFuture<ListAssociationsForLicenseConfigurationResponse, ListAssociationsForLicenseConfigurationError>

Lists the resource associations for a license configuration. Resource associations need not consume licenses from a license configuration. For example, an AMI or a stopped instance may not consume a license (depending on the license rules). Use this operation to find all resources associated with a license configuration.

fn list_license_configurations(
&self,
input: ListLicenseConfigurationsRequest
) -> RusotoFuture<ListLicenseConfigurationsResponse, ListLicenseConfigurationsError>

Lists license configuration objects for an account, each containing the name, description, license type, and other license terms modeled from a license agreement.

fn list_license_specifications_for_resource(
&self,
input: ListLicenseSpecificationsForResourceRequest
) -> RusotoFuture<ListLicenseSpecificationsForResourceResponse, ListLicenseSpecificationsForResourceError>

Returns the license configuration for a resource.

fn list_resource_inventory(
&self,
input: ListResourceInventoryRequest
) -> RusotoFuture<ListResourceInventoryResponse, ListResourceInventoryError>

Returns a detailed list of resources.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Lists tags attached to a resource.

fn list_usage_for_license_configuration(
&self,
input: ListUsageForLicenseConfigurationRequest
) -> RusotoFuture<ListUsageForLicenseConfigurationResponse, ListUsageForLicenseConfigurationError>

Lists all license usage records for a license configuration, displaying license consumption details by resource at a selected point in time. Use this action to audit the current license consumption for any license inventory and configuration.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<TagResourceResponse, TagResourceError>

Attach one of more tags to any resource.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<UntagResourceResponse, UntagResourceError>

Remove tags from a resource.

fn update_license_configuration(
&self,
input: UpdateLicenseConfigurationRequest
) -> RusotoFuture<UpdateLicenseConfigurationResponse, UpdateLicenseConfigurationError>

Modifies the attributes of an existing license configuration object. A license configuration is an abstraction of a customer license agreement that can be consumed and enforced by License Manager. Components include specifications for the license type (Instances, cores, sockets, VCPUs), tenancy (shared or Dedicated Host), host affinity (how long a VM is associated with a host), the number of licenses purchased and used.

fn update_license_specifications_for_resource(
&self,
input: UpdateLicenseSpecificationsForResourceRequest
) -> RusotoFuture<UpdateLicenseSpecificationsForResourceResponse, UpdateLicenseSpecificationsForResourceError>

Adds or removes license configurations for a specified AWS resource. This operation currently supports updating the license specifications of AMIs, instances, and hosts. Launch templates and AWS CloudFormation templates are not managed from this operation as those resources send the license configurations directly to a resource creation operation, such as RunInstances.

fn update_service_settings(
&self,
input: UpdateServiceSettingsRequest
) -> RusotoFuture<UpdateServiceSettingsResponse, UpdateServiceSettingsError>

Updates License Manager service settings.

Loading content...